+++ tests/BROKEN_test_space2_OLD.py
@@ -5,8 +5,11 @@ import haiku as hk
 from nndt.space2 import *
 
 
-class MyTestCase: # (unittest.TestCase):
+class MyTestCase:  # (BaseTestCase): # (unittest.TestCase):
 
+    # @classmethod
+    # def setUpClass(cls):
+    #     super().setUpClass()
     def setUp(self):
         self.name_list = os.listdir('./acdc_for_test')
         self.name_list = [filename[:10] for filename in self.name_list if 'patient' in filename]