+++ tests/test_haiku_modules.py
@@ -5,11 +5,16 @@ import jax
 import jax.numpy as jnp
 
 from nndt.haiku_modules import DescConv
-from nndt.haiku_modules import LipMLP
 from nndt.haiku_modules import LipLinear
+from nndt.haiku_modules import LipMLP
+from tests.base import BaseTestCase
 
 
-class HaikuModulesTestCase(unittest.TestCase):
+class HaikuModulesTestCase(BaseTestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
 
     def test_DescConv_init(self):
         def init(X):
@@ -76,6 +81,7 @@ class HaikuModulesTestCase(unittest.TestCase):
                            name="lip_mlp_%d" % 2,
                            activation=jax.nn.tanh)
             return ll(x)
+
         lLinear = hk.transform(init)
         self.assertIsInstance(lLinear, tuple)