+++ tests/test_import.py
@@ -1,7 +1,13 @@
 import unittest
 
+from tests.base import BaseTestCase
 
-class ImportTestCase(unittest.TestCase):
+
+class ImportTestCase(BaseTestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
 
     def test_math_core(self):
         from nndt import grid_in_cube, grid_in_cube2