+++ tests/test_math_core.py
@@ -2,9 +2,14 @@ import unittest
 
 from nndt.math_core import *
 from nndt.primitive_sdf import sdf_primitive_sphere
+from tests.base import BaseTestCase
 
 
-class MathCoreTestCase(unittest.TestCase):
+class MathCoreTestCase(BaseTestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
 
     def test_grid_in_cube(self):
         cube = grid_in_cube(spacing=(4, 4, 4), scale=4., center_shift=(2., 2., 2.))
@@ -193,7 +198,6 @@ class BarycentricGridTestCase(unittest.TestCase):
                                 [0.0, 0.0, 0.5, 0.5],
                                 [0.0, 0.0, 0.0, 1.0]]), coords)
 
-
     def test_rotation_matrix(self):
         self.assertTrue(jnp.allclose(rotation_matrix(0., 0., 0.), jnp.eye(3)))
         M = rotation_matrix(34., 34., 424.)
@@ -203,7 +207,6 @@ class BarycentricGridTestCase(unittest.TestCase):
         self.assertTrue(abs(float(jnp.linalg.norm(M[0])) - 1.0) < 0.0000001)
         self.assertTrue(abs(float(jnp.linalg.norm(M[1])) - 1.0) < 0.0000001)
         self.assertTrue(abs(float(jnp.linalg.norm(M[2])) - 1.0) < 0.0000001)
-
 
     def test_scale_xyz(self):
         xyz = jnp.array([1., 2., 3.])