+++ tests/test_vizualize.py
@@ -5,13 +5,18 @@ import unittest
 import numpy as np
 
 from nndt.vizualize import BasicVizualization
+from tests.base import BaseTestCase
 
 LOG_FOLDER = 'test_log'
 EXP_NAME = 'test_exp'
 EPOCHS = 100
 
 
-class VizualizeTestCase(unittest.TestCase):
+class VizualizeTestCase(BaseTestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
 
     def tearDown(self) -> None:
         shutil.rmtree(f'./{LOG_FOLDER}')